fix(server): emit one Y-up wire frame across all transports (#1841)
The Z-up -> Y-up axis swap was open-coded inline in each parquet serializer and
MISSING entirely from the JSON route, so `POST /api/v1/parse` (and the JSON SSE
stream) shipped raw IFC Z-up while `/parse/parquet*` shipped Y-up. A client that
correctly treats every transport alike therefore rendered JSON-loaded models
rotated. Same class of silent per-transport drift as the dropped `origin`.

- New `services::axis`: the ONE definition of the wire's axis convention
  (`zup_to_yup`, `zup_to_yup_f64`, `mesh_to_yup_in_place`), with unit tests
  pinning that positions, normals and origin rotate together.
- Both parquet serializers now call it (fused into their columnar extraction, so
  no extra pass); the JSON sync + SSE routes apply it to their meshes. The swap
  is not moved into `process_streaming`, which the parquet stream route shares.
- Documented on the client contract: server MeshData.positions are Y-up metres.

Module-size ratchet (shrink or split, never raise the budget):
- Split the geometry wire schema out of parquet.rs into parquet_schema.rs — the
  contract clients decode against now reads on its own. parquet.rs 566 -> 518
  (budget 522).
- Moved parquet_optimized.rs's inline test module to the ratchet-exempt sibling
  parquet_optimized_tests.rs, matching parquet_tests.rs. 683 -> 530 (budget 581).

Also refreshed scripts/api-surface.json for the three intentional new
server-client exports (ServerEntityIndex, ServerEntityColumns, EntityMetadata),
which is what failed the Node tests job.

Server: 68 cargo tests pass, clippy clean, ratchet green, api-surface matches.
