+++ apps/server/src/routes/parse/json.rs
@@ -7,6 +7,7 @@
 use super::cache_keys::{cache_symbolic_data, request_cache_key};
 use super::{extract_file, ParseQuery};
 use crate::error::ApiError;
+use crate::services::axis::mesh_to_yup_in_place;
 use crate::services::streaming::detect_schema_version;
 use crate::services::process_streaming;
 use crate::types::{MetadataResponse, ParseResponse, StreamEvent};
@@ -71,9 +72,19 @@ pub async fn parse_full(
     })
     .await?;
 
+    // Emit the SAME Y-up wire frame as the parquet transports (issue #1841).
+    // This route used to ship `result.meshes` verbatim — i.e. raw IFC Z-up —
+    // while `/parse/parquet*` swapped to Y-up server-side, so a client that
+    // (correctly) treats every transport alike rendered JSON-loaded models
+    // rotated. The frame now comes from one place: services::axis.
+    let mut meshes = result.meshes;
+    for mesh in &mut meshes {
+        mesh_to_yup_in_place(mesh);
+    }
+
     let response = ParseResponse {
         cache_key: cache_key.clone(),
-        meshes: result.meshes,
+        meshes,
         mesh_coordinate_space: result.mesh_coordinate_space,
         site_transform: result.site_transform,
         building_transform: result.building_transform,
@@ -131,7 +142,16 @@ pub async fn parse_stream(
         tessellation_quality,
         Some(admission_guard),
     )
-    .map(|event: StreamEvent| {
+    .map(|mut event: StreamEvent| {
+            // Same Y-up wire frame as every other transport (issue #1841) —
+            // `process_streaming` yields raw IFC Z-up meshes, and the parquet
+            // stream route applies the swap inside its serializer, so the JSON
+            // stream has to apply it here rather than in the shared producer.
+            if let StreamEvent::Batch { meshes, .. } = &mut event {
+                for mesh in meshes.iter_mut() {
+                    mesh_to_yup_in_place(mesh);
+                }
+            }
             let json = serde_json::to_string(&event).unwrap_or_else(|e| {
                 serde_json::to_string(&StreamEvent::Error {
                     message: e.to_string(),
