+++ apps/server/src/services/parquet_optimized.rs
@@ -13,6 +13,7 @@
 //!
 //! Typical additional compression: 3-5x over basic Parquet format.
 
+use crate::services::axis::{zup_to_yup, zup_to_yup_f64};
 use crate::types::MeshData;
 use arrow::array::{Float32Array, Float64Array, Int32Array, StringArray, UInt32Array, UInt8Array};
 use arrow::datatypes::{DataType, Field, Schema};
@@ -171,9 +172,10 @@ pub fn serialize_to_parquet_optimized(
         instance_ifc_types.push(&mesh.ifc_type);
         instance_mesh_indices.push(mesh_idx);
         instance_material_indices.push(material_idx);
-        instance_origin_x.push(mesh.origin[0]);
-        instance_origin_y.push(mesh.origin[2]);
-        instance_origin_z.push(-mesh.origin[1]);
+        let origin_yup = zup_to_yup_f64(mesh.origin);
+        instance_origin_x.push(origin_yup[0]);
+        instance_origin_y.push(origin_yup[1]);
+        instance_origin_z.push(origin_yup[2]);
         instance_geometry_class.push(mesh.geometry_class);
     }
 
@@ -236,19 +238,28 @@ pub fn serialize_to_parquet_optimized(
             );
         }
 
-        // Quantize and store vertices with Z-up to Y-up transform
-        // OPTIMIZATION: Apply coordinate transform server-side to eliminate client per-vertex loops
-        // IFC uses Z-up, WebGL uses Y-up. Transform: X stays same, new Y = old Z, new Z = -old Y
+        // Quantize and store vertices in the Y-up wire frame (services::axis),
+        // server-side so the client needs no per-vertex loop.
         for i in 0..vert_count {
-            vertex_x.push(quantize_position(mesh.positions[i * 3])); // X stays the same
-            vertex_y.push(quantize_position(mesh.positions[i * 3 + 2])); // New Y = old Z (vertical)
-            vertex_z.push(quantize_position(-mesh.positions[i * 3 + 1])); // New Z = -old Y (depth)
+            let (x, y, z) = zup_to_yup(
+                mesh.positions[i * 3],
+                mesh.positions[i * 3 + 1],
+                mesh.positions[i * 3 + 2],
+            );
+            vertex_x.push(quantize_position(x));
+            vertex_y.push(quantize_position(y));
+            vertex_z.push(quantize_position(z));
 
             if include_normals {
                 if mesh_has_normals {
-                    normal_x.push(mesh.normals[i * 3]); // X stays the same
-                    normal_y.push(mesh.normals[i * 3 + 2]); // New Y = old Z
-                    normal_z.push(-mesh.normals[i * 3 + 1]); // New Z = -old Y
+                    let (x, y, z) = zup_to_yup(
+                        mesh.normals[i * 3],
+                        mesh.normals[i * 3 + 1],
+                        mesh.normals[i * 3 + 2],
+                    );
+                    normal_x.push(x);
+                    normal_y.push(y);
+                    normal_z.push(z);
                 } else {
                     normal_x.push(0.0);
                     normal_y.push(0.0);
@@ -515,159 +526,5 @@ pub fn serialize_to_parquet_optimized_with_stats(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_optimized_parquet_serialization() {
-        // Create test data with some duplicate meshes
-        let wall_positions = vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0];
-        let wall_normals = vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0];
-        let wall_indices = vec![0, 1, 2];
-        let wall_color = [0.8, 0.8, 0.8, 1.0];
-
-        let meshes = vec![
-            // Two walls with same geometry (should be deduplicated)
-            MeshData::new(
-                1,
-                "IfcWall".to_string(),
-                wall_positions.clone(),
-                wall_normals.clone(),
-                wall_indices.clone(),
-                wall_color,
-            ),
-            MeshData::new(
-                2,
-                "IfcWall".to_string(),
-                wall_positions.clone(),
-                wall_normals.clone(),
-                wall_indices.clone(),
-                wall_color,
-            ),
-            // Different geometry
-            MeshData::new(
-                3,
-                "IfcSlab".to_string(),
-                vec![0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 2.0, 2.0, 0.0, 0.0, 2.0, 0.0],
-                vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
-                vec![0, 1, 2, 0, 2, 3],
-                [0.5, 0.5, 0.5, 1.0],
-            ),
-        ];
-
-        let (data, stats) = serialize_to_parquet_optimized_with_stats(&meshes, false).unwrap();
-
-        // Should deduplicate the two identical walls
-        assert_eq!(stats.input_meshes, 3);
-        assert_eq!(stats.unique_meshes, 2);
-        assert_eq!(stats.unique_materials, 2);
-        assert!(stats.mesh_reuse_ratio > 1.0);
-
-        // Should be very compact. Parquet has fixed per-column overhead, so
-        // tiny fixtures are dominated by it — the per-instance placement columns
-        // (origin_x/y/z + geometry_class, issue #1841) add four columns' worth of
-        // that fixed overhead, so the floor here is generous on purpose.
-        assert!(
-            data.len() < 8000,
-            "Expected compact output, got {} bytes",
-            data.len()
-        );
-    }
-
-    /// Contract test for issue #1841: the instance table MUST carry a
-    /// per-instance `origin` (Y-up) and `geometry_class`. Deduplication merges
-    /// bit-identical template geometry, so the ONLY thing that places a repeated
-    /// occurrence is its origin — dropping it collapses "N slabs into one slab"
-    /// at the template coordinates. Two identical slabs at different origins must
-    /// dedup to one mesh yet keep two distinct origins.
-    #[test]
-    fn instance_table_carries_origin_and_geometry_class() {
-        use arrow::array::{Float64Array, UInt8Array};
-        use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-
-        let slab = |id: u32, ifc_origin: [f64; 3]| {
-            MeshData::new(
-                id,
-                "IfcSlab".to_string(),
-                vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0],
-                vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
-                vec![0, 1, 2],
-                [0.5, 0.5, 0.5, 1.0],
-            )
-            .with_origin(ifc_origin)
-        };
-        // Same shape, two different placements → must dedup to ONE template.
-        let meshes = vec![
-            slab(1, [0.0, 0.0, 0.0]),
-            slab(2, [10.0, 20.0, 3.0]),
-        ];
-
-        let (data, stats) = serialize_to_parquet_optimized_with_stats(&meshes, false).unwrap();
-        assert_eq!(stats.unique_meshes, 1, "identical shapes must deduplicate");
-
-        // Unframe: [version:u8][flags:u8][instance_len:u32][...4 more lens][instance_parquet]...
-        let instance_len = u32::from_le_bytes(data[2..6].try_into().unwrap()) as usize;
-        let header = 2 + 5 * 4;
-        let instance_bytes = Bytes::copy_from_slice(&data[header..header + instance_len]);
-        let reader = ParquetRecordBatchReaderBuilder::try_new(instance_bytes)
-            .unwrap()
-            .build()
-            .unwrap();
-        let batch = reader.map(|b| b.unwrap()).next().unwrap();
-
-        let col = |name: &str| batch.schema().index_of(name).expect(name);
-        let oy = batch
-            .column(col("origin_y"))
-            .as_any()
-            .downcast_ref::<Float64Array>()
-            .unwrap();
-        let oz = batch
-            .column(col("origin_z"))
-            .as_any()
-            .downcast_ref::<Float64Array>()
-            .unwrap();
-        // geometry_class column must exist even when all-zero.
-        let _ = batch
-            .column(col("geometry_class"))
-            .as_any()
-            .downcast_ref::<UInt8Array>()
-            .unwrap();
-
-        assert_eq!(batch.num_rows(), 2, "both occurrences kept as instances");
-        // Instance 1: origin [10,20,3] IFC Z-up → Y-up [x, z, -y] = [10, 3, -20].
-        assert_eq!(oy.value(1), 3.0);
-        assert_eq!(oz.value(1), -20.0);
-    }
-
-    #[test]
-    fn test_quantization() {
-        assert_eq!(quantize_position(1.0), 10_000);
-        assert_eq!(quantize_position(0.0001), 1); // 0.1mm
-        assert_eq!(quantize_position(-1.5), -15_000);
-    }
-
-    #[test]
-    fn test_color_to_byte() {
-        assert_eq!(color_to_byte(0.0), 0);
-        assert_eq!(color_to_byte(1.0), 255);
-        assert_eq!(color_to_byte(0.5), 128);
-    }
-
-    /// Regression test for #586: meshes with positions but no normals
-    /// (e.g. `advanced_brep.ifc`) used to panic when `include_normals = true`.
-    #[test]
-    fn test_optimized_serialize_mesh_without_normals() {
-        let meshes = vec![MeshData::new(
-            42,
-            "IfcAdvancedBrep".to_string(),
-            vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0],
-            Vec::new(),
-            vec![0, 1, 2],
-            [0.8, 0.8, 0.8, 1.0],
-        )];
-
-        // Both code paths must survive empty normals.
-        assert!(serialize_to_parquet_optimized_with_stats(&meshes, false).is_ok());
-        assert!(serialize_to_parquet_optimized_with_stats(&meshes, true).is_ok());
-    }
-}
+#[path = "parquet_optimized_tests.rs"]
+mod optimized_tests;
