+++ packages/wasm/pkg/package.json
@@ -5,7 +5,7 @@
     "IFC-Lite Contributors"
   ],
   "description": "WebAssembly bindings for IFC-Lite",
-  "version": "4.1.0",
+  "version": "4.1.4",
   "license": "MPL-2.0",
   "repository": {
     "type": "git",
