+++ scripts/api-surface.json
@@ -3931,6 +3931,7 @@
     "CoordinateInfo: interface",
     "DataModel: interface",
     "DocumentAssociation: interface",
+    "EntityMetadata: interface",
     "ErrorResponse: interface",
     "GeometryDiagnostics: interface",
     "Georeferencing: interface",
@@ -3959,6 +3960,8 @@
     "Quantity: interface",
     "QuantitySet: interface",
     "ServerConfig: interface",
+    "ServerEntityColumns: interface",
+    "ServerEntityIndex: class",
     "StreamBatchEvent: interface",
     "StreamCompleteEvent: interface",
     "StreamErrorEvent: interface",
